fix(cli): fix URL encoding of bindPath and align with DocsWorkspace pattern
- Encode each segment of bindPath with encodeURIComponent before
  interpolating into the file-bind URL, resolving the Graphite comment
  amckinney flagged on PR #15308
- Refactor ThemeConfigProcessor to accept DocsWorkspace directly instead
  of a raw themeDir string, deriving the path internally
- Extract export logic into a ThemeExporter class that also takes
  DocsWorkspace directly, so v2 can adapt fern.yml and delegate
- Make exportDocsTheme and uploadDocsTheme thin CLI wrappers
